+++ src/org/moeaframework/core/comparator/LexicographicalComparator.java
@@ -17,7 +17,6 @@
  */
 package org.moeaframework.core.comparator;
 
-import java.io.Serializable;
 import java.util.Comparator;
 
 import org.moeaframework.core.Solution;
@@ -25,10 +24,7 @@ import org.moeaframework.core.Solution;
 /**
  * Compares solutions lexicographically.
  */
-public class LexicographicalComparator implements Comparator<Solution>,
-Serializable {
-
-	private static final long serialVersionUID = 2303639747960671103L;
+public class LexicographicalComparator implements Comparator<Solution> {
 
 	/**
 	 * Constructs a comparator for comparing solutions lexicographically.